Combined Frontend and Backend Login
Logging in and logging out officially works. Prevents accesses to pages unless a user from the DB is logged in. Checks to see if the inputted username and password is in the DB before logging in. If the user is logged in, the login button changes to logout and it removes the cookies of the user who was logged in. Jeniza fixed the page design being implemented and added the admin home page design with links. Hanna created the logos for both the navbar and the favicon.